Implements github suggestions - Hopefully less copypaste code.

+++ code/game/objects/structures/inflatable.dm
@@ -2,32 +2,30 @@
 	name = "inflatable"
 	w_class = 2
 	icon = 'icons/obj/inflatable.dmi'
+	var/deploy_path = null
+
+/obj/item/inflatable/attack_self(mob/user)
+	if(!deploy_path)
+		return
+	playsound(loc, 'sound/items/zip.ogg', 75, 1)
+	user << "<span class='notice'>You inflate \the [src].</span>"
+	var/obj/structure/inflatable/R = new deploy_path(user.loc)
+	src.transfer_fingerprints_to(R)
+	R.add_fingerprint(user)
+	qdel(src)
+
 
 /obj/item/inflatable/wall
 	name = "inflatable wall"
 	desc = "A folded membrane which rapidly expands into a large cubical shape on activation."
 	icon_state = "folded_wall"
-
-/obj/item/inflatable/wall/attack_self(mob/user)
-	playsound(loc, 'sound/items/zip.ogg', 75, 1)
-	user << "<span class='notice'>You inflate [src].</span>"
-	var/obj/structure/inflatable/wall/R = new /obj/structure/inflatable/wall(user.loc)
-	src.transfer_fingerprints_to(R)
-	R.add_fingerprint(user)
-	qdel(src)
+	deploy_path = /obj/structure/inflatable/wall
 
 /obj/item/inflatable/door/
 	name = "inflatable door"
 	desc = "A folded membrane which rapidly expands into a simple door on activation."
 	icon_state = "folded_door"
-
-/obj/item/inflatable/door/attack_self(mob/user)
-	playsound(loc, 'sound/items/zip.ogg', 75, 1)
-	user << "<span class='notice'>You inflate [src].</span>"
-	var/obj/structure/inflatable/door/R = new /obj/structure/inflatable/door(user.loc)
-	src.transfer_fingerprints_to(R)
-	R.add_fingerprint(user)
-	qdel(src)
+	deploy_path = /obj/structure/inflatable/door
 
 /obj/structure/inflatable
 	name = "inflatable"
@@ -38,11 +36,12 @@
 	icon = 'icons/obj/inflatable.dmi'
 	icon_state = "wall"
 
+	var/undeploy_path = null
 	var/health = 50.0
 
 /obj/structure/inflatable/wall
 	name = "inflatable wall"
-
+	undeploy_path = /obj/item/inflatable/wall
 
 /obj/structure/inflatable/New(location)
 	..()
@@ -85,8 +84,8 @@
 	deflate(1)
 
 /obj/structure/inflatable/attack_hand(mob/user as mob)
-		add_fingerprint(user)
-		return
+	add_fingerprint(user)
+	return
 
 /obj/structure/inflatable/attackby(obj/item/weapon/W as obj, mob/user as mob)
 	if(!istype(W)) return
@@ -114,10 +113,11 @@
 		src.transfer_fingerprints_to(R)
 		qdel(src)
 	else
-		//user << "<span class='notice'>You slowly deflate the inflatable wall.</span>"
-		visible_message("[src] slowly deflates.")
+		if(!undeploy_path)
+			return
+		visible_message("\The [src] slowly deflates.")
 		spawn(50)
-			var/obj/item/inflatable/R = new /obj/item/inflatable(loc)
+			var/obj/item/inflatable/R = new undeploy_path(src.loc)
 			src.transfer_fingerprints_to(R)
 			qdel(src)
 
@@ -149,6 +149,7 @@
 	opacity = 0
 
 	icon_state = "door_closed"
+	undeploy_path = /obj/item/inflatable/door
 
 	var/state = 0 //closed, 1 == open
 	var/isSwitchingStates = 0
